list connections after creating a new one
