+++ src/DbSchemas/DbSchemas/Program.cs
@@ -15,6 +15,7 @@ serviceCollection.AddScoped<ProgramDataService>();
 serviceCollection.AddScoped<DatabaseConnectionRecordService>();
 serviceCollection.AddScoped<DumpService>();
 serviceCollection.AddScoped<CliService>();
+serviceCollection.AddScoped<OutputService>();
 serviceCollection.AddScoped<DatabaseConnectionRecordRepository>();
 
 ServiceProvider serviceProvider = serviceCollection.BuildServiceProvider();
